Rename: rename common folder name
